+++ BidmadPluginSample/Assets/Plugins/iOS/Bidmad/BidmadSDK.framework/Headers/BIDMADBanner.h
@@ -49,7 +49,6 @@
 /// INITIALIZE ADS
 - (id)initWithParentViewController:(UIViewController *)parentVC adsPosition:(CGPoint)pointn bannerSize:(bannerSizeType) bannerTypeParam;
 - (id)initWithParentViewController:(UIViewController *)parentVC rootView:(UIView *)view    bannerSize:(bannerSizeType) bannerTypeParam;
-- (void)setTestDevice :(NSString *)deviceId;
 /// REQUEST ADS
 - (void)requestBannerView;
 /// DELETE ADS