+++ BidmadPluginSample/Assets/Plugins/iOS/Bidmad/BidmadSDK.framework/Headers/BIDMADInterstitial.h
@@ -93,8 +93,6 @@
 
 - (void)removeInterstitialADS;
 
-- (void)setTestDevice :(NSString *)deviceId;
-
 - (void)sendLog :(NSDictionary *) info :(NSString *) advertisementType :(NSString *) logType;
 
 @end