+++ BidmadPluginSample/Assets/Plugins/iOS/Bidmad/BidmadSDK.framework/Headers/BIDMADRewardVideo.h
@@ -86,8 +86,6 @@
 
 - (void) sendLog :(NSDictionary *) info :(NSString *) advertisementType :(NSString *) logType;
 
-- (void) setTestDevice :(NSString *)deviceId;
-
 - (void) releaseBidmadResource;