+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 2.7.0
+#### Changes
+- GDPR Function Added (for Google Ad Networks)
+- [iOS] GDPR Function Fix (for Non-Google Ad Networks)
+#### Integration Bidmad SDK version
+- Android : bidmad-androidx-g 1.1.0
+- iOS : BidmadSDK.framework 2.6.0
+
 # Version 2.6.1
 #### Changes
 - Fix Bugs